Rename card.php and pipeline_button.php to be consistent with the classes they contain.
